Synchronous SVG mipmaps
